pyproject.toml
[tool.poetry]
name = "socoal_network_engagement_bot"
version = "0.1.0"
description = " Python-based simple RESTful API that tracks and provides engagement insights for social media accounts (e.g., Twitter or Instagram) based on user activity"
authors = ["Mohsen Baghdadi  <mohsen.baghdadi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
alembic = "^1.14.0"
pydantic = "^2.10.3"
pydantic-settings = "^2.7.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
geocoder = "^1.38.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
tenacity = "^9.0.0"
python-jose = "^3.3.0"
redis = "^5.2.1"
passlib = "^1.7.4"
ua-parser = "^1.0.0"
aioredis = "^2.0.1"
aiofiles = "^24.1.0"
aiohttp = "^3.11.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
